--- create_app/core/management/templates/project_template/manage.py-tpl
@@ -1,13 +1,0 @@
-from flask.cli import FlaskGroup
-import click
-import os
-
-from app import create_app
-
-@click.group(cls=FlaskGroup, create_app=create_app)
-@click.pass_context
-def cli(ctx):
-    os.environ['FLASK_ENV'] ='development'
-
-if __name__ == "__main__":
-    cli()
