+++ flask_create_app/core/management/templates/project_template/app/local_settings.py-tpl
