+++ flask_create_app/core/management/templates/project_template/app/models/user_model.py-tpl
@@ -1,0 +1,47 @@
+# Copyright 2014 SolidBuilds.com. All rights reserved
+#
+# Authors: Ling Thio <ling.thio@gmail.com>
+
+from flask_user import UserMixin
+# from flask_user.forms import RegisterForm
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, validators
+from app import db
+
+
+# Define the User data model. Make sure to add the flask_user.UserMixin !!
+class User(db.Model, UserMixin):
+    __tablename__ = 'users'
+
+    id = db.Column(db.Integer, primary_key=True)
+    public_id = db.Column(db.String(50), unique=True)
+    username = db.Column(db.String(50), nullable=True)
+    email = db.Column(db.String(120), nullable=False, unique=True)
+    password = db.Column(db.String(80), nullable=False)
+    is_active = db.Column(db.Boolean, default=True)
+    is_admin = db.Column(db.Boolean, default=False)
+    roles = db.relationship('Role', secondary='users_roles',
+                            backref=db.backref('users', lazy='dynamic'))
+
+
+# Define the Role data model
+class Role(db.Model):
+    __tablename__ = 'roles'
+    id = db.Column(db.Integer(), primary_key=True)
+    name = db.Column(db.String(50), nullable=False, unique=True)  # for @roles_accepted()
+
+# Define the UserRoles association model
+class UsersRoles(db.Model):
+    __tablename__ = 'users_roles'
+    id = db.Column(db.Integer(), primary_key=True)
+    user_id = db.Column(db.Integer(), db.ForeignKey('users.id', ondelete='CASCADE'))
+    role_id = db.Column(db.Integer(), db.ForeignKey('roles.id', ondelete='CASCADE'))
+
+
+# Define the User form
+class UserForm(FlaskForm):
+    username = StringField('Username', validators=[
+        validators.DataRequired('First name is required')])
+    email = StringField('Last name', validators=[
+        validators.DataRequired('Last name is required')])
+    submit = SubmitField('Save')
