+++ flask_create_app/core/management/templates/project_template/app/settings.py-tpl
