+++ flask_create_app/core/management/templates/project_template/app/static/style.css
