+++ flask_create_app/core/management/templates/project_template/app/templates/account/register.html
