+++ flask_create_app/core/management/templates/project_template/app/templates/base.html
@@ -20,7 +20,6 @@
                 <div class="collapse navbar-collapse" id="navbarToggle">
                     <div class="navbar-nav mr-auto">
                     <a class="nav-item nav-link" href="/">Home</a>
-                    <a class="nav-item nav-link" href="/about">About</a>
                     </div>
                     <!-- Navbar Right Side -->
                     <div class="navbar-nav">
