+++ flask_create_app/core/management/templates/project_template/app/templates/index.html
@@ -4,7 +4,7 @@
 {% block content %}
 <div class="container">
     <div  style="text-align: center;">
-        <h3>Flask Project created</h3>
+        <h3>Flask Project</h3>
     </div>
     <div style="text-align: left;">
         <h5>Project has following directory structure</h5>
@@ -19,7 +19,9 @@
                 <li>templates
                     <p  class="text-muted">html content</p>
                 </li>
-                <li>views</li>
+                <li>views
+                    <p  class="text-muted">view layer</p>
+                </li>
 
             </ul>
             <li>tests</li>
