+++ flask_create_app/core/management/templates/project_template/app/views/__init__.py-tpl
