+++ flask_create_app/core/management/templates/project_template/requirements.txt
@@ -7,3 +7,4 @@ Flask-Script==2.0.6
 Flask-SQLAlchemy==2.4.0
 Flask-User==1.0.1.5
 Flask-WTF==0.14.2
+flask_bcrypt
