+++ flask_create_app/core/management/templates/project_template/run.py-tpl
